XD-111 unit tests for groupByOpPanelModel.ts

+++ ts/components/dag/DagSubGraph.ts
@@ -116,10 +116,12 @@ class DagSubGraph extends DagGraph {
             } else {
                 args = xcHelper.getXcalarInputFromNode(queryNode);
             }
-            let nodeId: DagNodeId = this._nameIdMap[args.dest];
-            let node: DagNode = this.getNode(nodeId);
-            if (node != null) { // could be a drop table node
-                node.beRunningState();
+            if (queryNode.operation !== XcalarApisTStr[XcalarApisT.XcalarApiDeleteObjects]) {
+                let nodeId: DagNodeId = this._nameIdMap[args.dest];
+                let node: DagNode = this.getNode(nodeId);
+                if (node != null) { // could be a drop table node
+                    node.beRunningState();
+                }
             }
         });
     }
